backend/bridge/requirements.txt
websockets==12.0
watchdog==4.0.0
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
